error msg now to stderr and changed atoi to strtol

+++ exercise_1A_1B/1B/circular_buffer.h
@@ -29,8 +29,8 @@ int shm_fd = 0;
 // define structs
 struct edge
 {
-    int vertex_u;
-    int vertex_v;
+    long vertex_u;
+    long vertex_v;
 };
 
 struct element
